# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "React Online Store",
        "version": "0.1"
    },
    "paths": {
...
        "/api/admin/categories": {
            "get": {
                "tags": [
                    "AdminCategory"
                ],
                "summary": "Display a listing of the Categories.",
                "operationId": "71fce704423c89eab31adce4a23df70b",
                "responses": {
                    "200": {
                        "description": "Display a listing of the Categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "AdminCategory"
                ],
                "summary": "Store a newly created Category.",
                "operationId": "810e8b9fee1619d4229208cb942c486a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Stored a newly created Category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/{id}": {
            "get": {
                "tags": [
                    "AdminCategory"
                ],
                "summary": "Display the specified Category by ID.",
                "operationId": "dd1a3fd24575b2b12b6c8235fd29de4e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified Category by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "AdminCategory"
                ],
                "summary": "Update the specified Category by ID.",
                "operationId": "6b267ce72356fa2805c1484f48318133",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Updated the specified Category by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "AdminCategory"
                ],
                "summary": "Remove the specified Category by ID.",
                "operationId": "6ee6ff59b8676a6da057579bba1331ad",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed the specified Category by ID"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/restore": {
            "post": {
                "tags": [
                    "AdminCategory"
                ],
                "summary": "Restore the specified Categories by ID.",
                "operationId": "0fec3a7c4071f4101d4cd674637a5034",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RestoreCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restore categories",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/CategoryResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/categories/{id}/force": {
            "delete": {
                "tags": [
                    "AdminCategory"
                ],
                "summary": "Remove the specified Category by ID.",
                "operationId": "6b9e29be92a4391fe3b8a7c1d03e4ee7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Removed the specified Category by ID"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Display the specified Category by ID.",
                "operationId": "09767bb2e47dde77dd1e2591428b0ae1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Display the specified Category by ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoryResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RestoreCategoryRequest": {
                "required": [
                    "id"
                ],
                "properties": {
                    "id": {
                        "description": "Categories list for restore",
                        "type": "array",
                        "items": {
                            "type": "integer"
                        }
                    }
                },
                "type": "object"
            },
            "StoreCategoryRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Category name",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
            "UpdateCategoryRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "Category name",
                        "type": "string",
                        "maxLength": 255
                    }
                },
                "type": "object"
            },
...
            "CategoryResource": {
                "properties": {
                    "id": {
                        "description": "Class CategoryResource",
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "image": {
                        "$ref": "#/components/schemas/ImageResource"
                    }
                },
                "type": "object"
            },
            "ImageResource": {
                "properties": {
                    "id": {
                        "description": "Class ImageResource",
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "origin_name": {
                        "type": "string"
                    },
                    "width": {
                        "type": "string"
                    },
                    "height": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "jwt",
                "scheme": "bearer"
            }
        }
    }
}